latest update (2/05/2023)

+++ app/Http/Controllers/SecretaryController.php
@@ -5,7 +5,11 @@ namespace App\Http\Controllers;
 use App\Models\Addtocartservice;
 use App\Models\Admin;
 use App\Models\Appointment;
+use App\Models\AuditTrail;
+use App\Models\BusinessHour;
 use App\Models\Discount;
+use App\Models\Modeofpayment;
+use App\Models\Reservationfee;
 use App\Models\Service;
 use App\Models\Transaction;
 use App\Models\User;
@@ -16,6 +20,7 @@ use Illuminate\Validation\Rule;
 use Illuminate\Support\Facades\Validator;
 use Carbon\Carbon;
 use DataTables;
+use Illuminate\Support\Facades\File;
 
 
 class SecretaryController extends Controller
@@ -88,6 +93,31 @@ class SecretaryController extends Controller
             //                               ->with('datas', $gender_records)
             //                              ->with('appointments', $appointments)
             //                               ;
+
+            $transaction = Transaction::select(DB::raw('SUM(total) as total'))
+            ->whereYear('created_at', date('Y'))
+            ->groupBy(DB::raw('MONTH(created_at)'))
+            ->pluck('total');
+        
+            
+        $transactionArray = $transaction->map(function ($item) {
+            return (string) $item;
+        })->toArray();
+
+
+            $try = DB::select(DB::raw('
+    SELECT SUM(total) as total, MONTH(created_at) as month FROM (
+        SELECT total, created_at FROM transactions WHERE YEAR(created_at) = YEAR(NOW())
+        UNION ALL
+        SELECT reservation_fee, created_at FROM appointments WHERE YEAR(created_at) = YEAR(NOW())
+    ) as combined_totals
+    GROUP BY MONTH(created_at)
+    ORDER BY MONTH(created_at)
+'));
+
+
+
+$totals = array_column($try, 'total', 'month');
      
             $users = User::all()->count();
             $pending = Appointment::where('status', 'Pending')->count();
@@ -99,8 +129,8 @@ class SecretaryController extends Controller
             $name= auth()->user()->fname;
     
             // $patient = User::select('fname')->distinct()->get();
-            return view('secretary.dashboard') ->with('name', $name)
-                                        //   ->with('patients', $patient)
+            return view('secretary.dashboard', compact('totals', 'transactionArray')) ->with('name', $name)
+                                        // ->with('patients', $patient)
                                           ->with('users', $users)
                                           ->with('pending', $pending)
                                           ->with('transaction', $totalsales)
@@ -225,8 +255,8 @@ class SecretaryController extends Controller
 
     //service
     public function service_show(){
-        $service = Service::all();
-        return view('/secretary/service', ['services' => $service]);
+        $service = DB::table('services')->paginate(8);
+        return view('secretary.system_settings.service', ['services' => $service]);
     }
 
 public function fetch_service(){
@@ -340,7 +370,6 @@ public function fetch_service(){
 
 public function view_queuing(Request $request){
 
-
     if ($request->ajax()) {
         $data = DB::table('appointments')->where('status', 'pending')->whereDate('date', '=', date('Y-m-d'))->whereDate('time', '>', date('H:i:s'))
         ->orderBy('time', 'asc');
@@ -362,8 +391,6 @@ public function view_queuing(Request $request){
                 ->make(true);
     }
     return view('secretary.queuing');
-
-
 
 }
 
@@ -459,31 +486,41 @@ public function store_addtocart(Request $request){
 
 }
 public function store_billing(Request $request){
-    $sum = Addtocartservice::sum('price');
-    // $insert = Addtocartservice::where('billing_no', $request->billingno)->update(['sub_total' => $sum])
+    if (Addtocartservice::count() == 0) {
+        return response()->json(['status' => '400']);
+    }else{
+        $sum = Addtocartservice::sum('price');
+        // $insert = Addtocartservice::where('billing_no', $request->billingno)->update(['sub_total' => $sum])
+        $addtocart = Addtocartservice::all();
+        foreach ($addtocart as $data) {
 
-    $addtocart = Addtocartservice::all();
+            $billing = new Transaction();
+            $billing->transno = $data->transno;
+            $billing->user_id = $data->user_id;
+            $billing->fullname = $data->fullname;
+            $billing->servicecode = $data->servicecode;
+            $billing->service = $data->service;
+            $billing->price = $data->price;
+            $billing->sub_total = $sum;
+            $billing->total = $sum;
+            $billing->status = 'Pending';
+            $billing->save();
+        }
+        Addtocartservice::truncate();  
 
-    foreach ($addtocart as $data) {
+        $audit_trail = new AuditTrail();
+        $audit_trail->user_id = Auth::user()->id;
+        $audit_trail->username = Auth::user()->username;
+        $audit_trail->activity = 'Create transaction';
+        $audit_trail->usertype = Auth::user()->usertype;
+        $audit_trail->save();
 
-        $billing = new Transaction();
-        $billing->transno = $data->transno;
-        $billing->user_id = $data->user_id;
-        $billing->fullname = $data->fullname;
-        $billing->servicecode = $data->servicecode;
-        $billing->service = $data->service;
-        $billing->price = $data->price;
-        $billing->sub_total = $sum;
-        $billing->total = $sum;
-        $billing->status = 'Pending';
-        $billing->save();
+        return response()->json([
+           'status' => 200,
+           'message' => 'Saved successfully',
+        ]);
+
     }
-    Addtocartservice::truncate();  
-
-    return response()->json([
-       'status' => 200,
-       'message' => 'Saved successfully',
-    ]);
 
 }
 
@@ -517,15 +554,15 @@ public function index_billing(Request $request){
                     if($row->status == "Pending"){
                         
                          $btn = '<button type="button" data-id="' . $row->transno . '" class="payment btn  btn-success btn-sm">Pay now</button>';
-                        $btn = $btn.'<a href="/admin/billing/viewBilling/ ' . $row->transno . ' " class="btn btn-primary btn-sm" style="margin-left:5px"    >View</a>';
-                        $btn = $btn.' <a href="/admin/billing/editBilling/' . $row->transno . '" class="btn btn-danger btn-sm">Delete</a>';
+                        $btn = $btn.'<a href="/secretary/billing/viewBilling/ ' . $row->transno . ' " class="btn btn-primary btn-sm" style="margin-left:5px"    >View</a>';
+                        $btn = $btn.'<button  style="margin-left:5px" class="delete btn btn-sm btn-danger" data-id="' . $row->transno . '">Delete</button>';
                         $size = '<div style="width: 200px">' . $btn . '</div>';                
                         return $size;
                        
                     }else{
 
-                        $btn = ' <a href="/admin/billing/viewBilling/' . $row->transno . '" class="btn btn-primary btn-sm">View</a>';
-                        $btn = $btn.'  <a href="/admin/billing/editBilling/' . $row->transno . '" class="btn btn-danger btn-sm">Delete</a>';
+                        $btn = ' <a href="/secretary/billing/viewBilling/' . $row->transno . '" class="btn btn-primary btn-sm">View</a>';
+                        $btn = $btn.'<button  style="margin-left:5px" class="delete btn btn-sm btn-danger" data-id="' . $row->transno . '">Delete</button>';
                 $size = '<div style="width: 200px">' . $btn . '</div>';                
                         return $size;
 
@@ -538,11 +575,12 @@ public function index_billing(Request $request){
     }
 //    $billing =  DB::table('transactions')->distinct()->select('transno', 'user_id', 'fullname', 'sub_total', 'status', 'total' )->orderBy('transno', 'desc')->paginate(10, ['*'], 'addtocart');
     $discount = Discount::all();
+    $mops = Modeofpayment::all();
     // $billing = Transaction::distinct()->select('transno', 'user_id', 'fullname', 'sub_total', 'status', 'total' )->orderBy('transno', 'desc')->get();
 
     return view('secretary.billing', [
                                     'discounts' =>$discount,
-                          
+                                    'mops' =>$mops,
                                   ]);
 }
 
@@ -565,7 +603,7 @@ public function addtocart_getalldata($id){
         //------------------- discount ---------------------------//
         public function discount_show(){
             $data = Discount::all();
-            return view('system_settings.discount', ['discounts' => $data]);
+            return view('secretary.system_settings.discount', ['discounts' => $data]);
         }
     
         public function fetch_discount(){
@@ -601,6 +639,14 @@ public function addtocart_getalldata($id){
                 $discount->discountname = $request->input('discountname');
                 $discount->percentage = $request->input('percentage');
                 $discount->save();
+
+                $audit_trail = new AuditTrail();
+                $audit_trail->user_id = Auth::user()->id;
+                $audit_trail->username = Auth::user()->username;
+                $audit_trail->activity = 'Create new discount';
+                $audit_trail->usertype = Auth::user()->usertype;
+                $audit_trail->save();
+
                 return response()->json([
                     'status'=>200,
                     'message' => 'discount added successfully',
@@ -643,7 +689,14 @@ public function addtocart_getalldata($id){
                 );
     
                 DB::table('discounts')->where('discountcode', $discountcode)->update($arrItem);
-    
+                
+                $audit_trail = new AuditTrail();
+                $audit_trail->user_id = Auth::user()->id;
+                $audit_trail->username = Auth::user()->username;
+                $audit_trail->activity = 'Update discount';
+                $audit_trail->usertype = Auth::user()->usertype;
+                $audit_trail->save();
+
                 return response()->json([
                     'status'=>200,
                         'message' => 'discount updated successfully',
@@ -656,13 +709,163 @@ public function addtocart_getalldata($id){
         public function delete_discount($discountcode)
         {   
             DB::table('discounts')->where('discountcode', $discountcode)->delete();
-    
+            
+            $audit_trail = new AuditTrail();
+            $audit_trail->user_id = Auth::user()->id;
+            $audit_trail->username = Auth::user()->username;
+            $audit_trail->activity = 'Delete';
+            $audit_trail->usertype = Auth::user()->usertype;
+            $audit_trail->save();
+
             return response()->json([
                 'status'=>200,
                     'message' => 'deleted successfully',
                 
             ]);
         }
+
+
+        public function view_billing($id){
+            $infos = Transaction::with('user')->where('transno', $id)->first();
+    
+            $services = Transaction::where('transno', $id)->get();
+          
+            return view('secretary.viewBilling', compact('services', 'infos'));
+        }
+
+        public function show_businesshours(){
+            $hours = BusinessHour::where('day', 'Monday')->whereNotIn('from', ['23:59:00'])->orderBy('from', 'asc')->get();
+            
+            $day = BusinessHour::where('day', 'Monday')->select('day', 'off')->distinct()->get();
+        return view('secretary.system_settings.businesshours', ['hours' =>$hours, "days" => $day]);
+     }
+
+     public function index_modeofpayment(){
+        $mops = DB::table('modeofpayments')->orderBy('created_at', 'desc')->paginate(5);
+
+        return view('secretary.system_settings.modeofpayment', compact('mops'));
+    }
+
+
+    public function index_reservationfee_setting(){
+        $reservationfee = Reservationfee::first();
+        return view('secretary.system_settings.reservationfee', compact('reservationfee'));
+    }
+
+
+
+    public function index_myprofile(){
+        return view('secretary.profile.index');
+    }
+
+    public function edit_myprofile(){
+        return view('secretary.profile.edit');
+    }
+
+    public function update_myprofile(Request $request){
+        $validated = $request->validate([
+            "first_name" => ['required'],
+            "mname" => [''],
+            "last_name" => ['required',],
+            "birthday" => ['required'],
+            "age" => ['required'],
+            "address" => ['required'],
+            "gender" => ['required'],
+        ],[
+          'first_name.required' => 'First name is required',
+          'last_name.required' => 'Last name is required',
+          'birthday.required' => 'Birthday is required',
+          'age.required' => 'Age is required',
+          'address.required' => 'Address is required',
+          'gender.required' => 'gender is required',
+        ]);
+
+    $user = User::where('id', Auth::user()->id)->first();
+    $input = $request->all();
+    
+    $user->fname = $input['first_name'];
+    $user->mname = $input['mname'];
+    $user->lname = $input['last_name'];
+    $user->birthday = $input['birthday'];
+    $user->address = $input['address'];
+    $user->age= $input['age'];
+    $user->gender = $input['gender'];
+
+    $user->save();
+
+    
+    $audit_trail = new AuditTrail();
+    $audit_trail->user_id = Auth::user()->id;
+    $audit_trail->username = Auth::user()->username;
+    $audit_trail->activity = 'Update profole';
+    $audit_trail->usertype = Auth::user()->usertype;
+    $audit_trail->save();
+
+    return redirect('/secretary/myprofile')->with('success', 'Updated successfully');
+
+    }
+
+    public function update_profile_pic($id, Request $request){
+        $input = $request->all();
+        $validated = $request->validate([
+            "profilepic" => 'required|mimes:png,jpg,jpeg',
+ 
+        ],[
+            "profilepic" =>'The picture must be a file of type: png, jpg, jpeg.'
+        ]);
+
+        $user = User::where('id', Auth::user()->id)->first();
+        $path = public_path('profilepic/'.$user->profile_pic);
+
+        if(File::exists($path)){
+            File::delete($path);
+        }
+        $filename = date('YmdHis'). '.' . $input['profilepic']->getClientOriginalExtension();
+        $input['profilepic']->move(public_path('profilepic/'), $filename) ;
+        $input['profilepic'] = $filename;
+        $user->profile_pic = $filename;
+
+        $user->save();
+
+
+        
+        $audit_trail = new AuditTrail();
+        $audit_trail->user_id = Auth::user()->id;
+        $audit_trail->username = Auth::user()->username;
+        $audit_trail->activity = 'Change password';
+        $audit_trail->usertype = Auth::user()->usertype;
+        $audit_trail->save();
+       return redirect()->back();
+    }
+
+    public function index_changepass(){
+        return view('secretary.profile.changepass');
+    }
+
+    public function update_changepass(Request $request){
+
+   $input = $request->all();
+        
+    if(password_verify($input['oldpassword'], Auth::user()->password)){
+        $validated = $request->validate([
+            "password" => 'required|confirmed|min:8',
+        ],[
+          'password.required' => 'Password is required',
+          'password.confirmed' => 'Password did not match',
+        ]);
+
+        $user = User::where('id', Auth::user()->id)->first();
+        $encrypt = bcrypt($request->input('password'));
+        $user->password = $encrypt;
+        $user->save();
+        return redirect()->back()->with('success', 'Updated successfully');
+         
+    }else{
+        return redirect()->back()->with('oldpassword', 'The password did not match with the current password.');
+    }
+
+}
+
 
 
 }
